The POD search widget is created
